+++ craftai/errors.py
@@ -57,7 +57,6 @@ class CraftAiNotFoundError(CraftAiError):
 class CraftAiDecisionError(CraftAiError):
   """Raised when some issue is encountered when trying to find a decision"""
   def __init__(self, message):
-    print("message", message)
     self.message = "".join(("Unable to take decision, ", message))
     super(CraftAiDecisionError, self).__init__(message)
 
@@ -65,7 +64,7 @@ class CraftAiDecisionError(CraftAiError):
 class CraftAiNullDecisionError(CraftAiDecisionError):
   """Raised when some issue is encountered when trying to find a decision"""
   def __init__(self, message):
-    self.message = "".join(("Unable to take decision: ", message))
+    self.message = "".join(("Unable to take decision, ", message))
     super(CraftAiNullDecisionError, self).__init__(message)