+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/craft-ai/craft-ai-client-python/compare/v1.6.0...HEAD) ##
-- Change error messages text to fit interpreter messages.
+### Fixed ###
+- Fix interpreter test suite to check message expectation (check issue [here](https://github.com/craft-ai/craft-ai-client-python/issues/59)). Some error messages have been updated to fit the interpreter expectation.
 
 ## [1.6.0](https://github.com/craft-ai/craft-ai-client-python/compare/v1.5.0...v1.6.0) - 2017-08-22 ##
 ### Added ###